fix(server): clear pre-existing clippy -D warnings findings
The branch failed `cargo clippy --all-targets -- -D warnings` (server-ci gate)
on 7 findings that predate this work: collapsible if-lets (ai/client, ai/routes),
an unnecessary u64→u64 cast (cloud/reader), items after the test module
(catalog/reconcile — moved spawn() above it), and deprecated GenericArray
from_slice in crypto (aes-gcm nonce built via From<[u8;12]>/try_into instead —
no generic-array bump needed). Added crypto round-trip tests (there were none)
to prove the nonce change keeps the AEAD seal/open intact.

clippy --all-targets -D warnings now clean; full test suite green (24 passed).

+++ server/src/cloud/reader.rs
@@ -34,7 +34,7 @@ pub async fn list_files(
             Ok(meta) => {
                 entries.push(FileEntry {
                     path: meta.location.to_string(),
-                    size: meta.size as u64,
+                    size: meta.size,
                     last_modified: Some(meta.last_modified.to_string()),
                 });
             }
